Added basic type checking in api arguments, updated tests
